chore: Create common factory for enum (#666)
Added spec for enum and created common factory for processing parameter and property type enums.
