+++ examples/spec/twilio_enum_v1.yaml
@@ -1,0 +1,106 @@
+info:
+  contact:
+    email: support@twilio.com
+    name: Twilio Support
+    url: https://support.twilio.com
+  description: This is the public Twilio REST API.
+  license:
+    name: Apache 2.0
+    url: https://www.apache.org/licenses/LICENSE-2.0.html
+  termsOfService: https://www.twilio.com/legal/tos
+  title: Twilio - CONTENT
+  version: 1.11.0
+openapi: 3.0.1
+
+paths:
+  /v1/content:
+    servers:
+      - url: https://enum.twilio.com
+    get:
+      summary: Retrieve items with various filter options
+      parameters:
+        # PARAMETER_SINGLE
+        - name: singleParam
+          in: query
+          description: A single enum value as a query parameter
+          required: false
+          schema:
+            type: string
+            enum: 
+              - asc
+              - desc
+            example: asc
+
+        # PARAMETER_ARRAY
+        - name: arrayParam
+          in: query
+          description: Order items using an array of enums
+          required: false
+          schema:
+            type: array
+            items:
+              type: string
+              enum:
+                - asc
+                - desc
+            example: [asc]
+
+        # PARAMETER_ARRAY
+        - name: arrayParamRef
+          in: query
+          required: false
+          description: An array parameter referencing a reusable schema
+          schema:
+            type: array
+            items:
+              $ref: '#/components/schemas/singleReusable'
+
+        - name: singleParamRef
+          in: header
+          schema:
+            $ref: '#/components/schemas/singleReusable'
+
+      responses:
+        '200':
+          description: A list of items
+          content:
+            application/json:
+              schema:
+                type: array
+                items:
+                  $ref: '#/components/schemas/Item'
+
+components:
+  schemas:
+    Item:
+      type: object
+      properties:
+        # PROPERTY_SINGLE
+        singleProperty:
+          type: string
+          enum: [available, pending, sold]
+          description: A single enum value as a property in a schema
+          example: available
+
+        # PROPERTY_ARRAY
+        arrayProperty:
+          type: array
+          items:
+            type: string
+            enum: [new, sale, featured]
+          description: An array of enum values in a schema property
+          example: [new, featured]
+
+        # REUSABLE_ARRAY
+        arrayReusable:
+          type: array
+          items:
+            $ref: '#/components/schemas/singleReusable'
+          description: An array of reusable enum defined in components
+          example: [electronics, clothing]
+
+  # REUSABLE_SINGLE
+    singleReusable:
+      type: string
+      enum: [electronics, furniture, clothing]
+      description: A reusable single-value enum defined in components
